test: move filter test ATTRIBUTES to conftest.py as fixture
